Use Lookup for menu file/preset lookup
An array of all menu/preset file paths was filtered in order to look 0
or more results from the menu/preset filename. A lookup will do the
exact same thing but on initialization. This simplifies lookup by just
having to provide the key without any transformations and improves
performance because the keys and values have already been grouped.